+++ src-dynamic-route/app/app.routes.ts
@@ -1,0 +1,10 @@
+import { ModuleWithProviders } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { HomeComponent } from './shared/components/home/home.component';
+
+const routes: Routes = [
+    { path: '', component: HomeComponent }
+];
+
+export const routing: ModuleWithProviders = RouterModule.forRoot(routes);